Minor refactor
- Move switch statement to EnemyOperationsController
- Update namespaces to match filepath (except for Enemy, which is still outside of AI)
