+++ TAC_AI/Enemy/EnemyOperations/RWheeled.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TAC_AI.AI.Enemy
+namespace TAC_AI.AI.Enemy.EnemyOperations
 {
-    public class RWheeled
+    public static class RWheeled
     {
-        public static void TryAttack(AIECore.TankAIHelper thisInst, Tank tank, RCore.EnemyMind mind)
+        public static void TryAttack(AIECore.TankAIHelper thisInst, Tank tank, EnemyMind mind)
         {
             BGeneral.ResetValues(thisInst);
             thisInst.Attempt3DNavi = false;
